Read and write SPSS files

read_sav() reads both .sav and .zsav files. read_por_declared() reads .por files. read_spss_declared() uses either read_por_declared() or read_sav_declared() based on the file extension.

Usage

read_sav_declared(
  file,
  encoding = NULL,
  user_na = FALSE,
  col_select = NULL,
  skip = 0,
  n_max = Inf,
  .name_repair = "unique"
)

read_por_declared(
  file,
  user_na = FALSE,
  col_select = NULL,
  skip = 0,
  n_max = Inf,
  .name_repair = "unique"
)

read_spss_declared(
  file,
  user_na = FALSE,
  col_select = NULL,
  skip = 0,
  n_max = Inf,
  .name_repair = "unique"
)

Arguments

file

The file to read.

encoding

The character encoding used for the file. The default, NULL, use the encoding specified in the file, but sometimes this value is incorrect and it is useful to be able to override it.

user_na

If TRUE variables with user defined missing will be read into labelled_spss() objects. If FALSE, the default, user-defined missings will be converted to NA.

col_select

One or more selection expressions, like in dplyr::select(). Use c() or list() to use more than one expression. See ?dplyr::select for details on available selection options. Only the specified columns will be read from data_file.

skip

Number of lines to skip before reading data.

n_max

Maximum number of lines to read.

Value

A tibble, data frame variant with nice defaults.

Variable labels are stored in the "label" attribute of each variable. It is not printed on the console, but the RStudio viewer will show it.

Details

Currently haven can read and write logical, integer, numeric, character and factors. This is not ideal for social science surveys, and therefore we created a wrapper haven's functions to read the data into declared types.

Examples

path <- system.file("examples", "ZA5688_v6_sample.sav", package = "surveydataset")
read_sav_declared(path)
#> Error: '' does not exist in current working directory ('/home/runner/work/surveydataset/surveydataset/docs/reference').

tmp <- tempfile(fileext = ".sav")
read_sav_declared(tmp)
#> Error: '/tmp/Rtmp1RBDTz/file298e7130560b.sav' does not exist.