Create a survey

Create a dataset suitable for working with social sciences surveyd data.

Usage

survey(
  x,
  Dimensions,
  Measures,
  Attributes,
  Constants,
  Title = NULL,
  Identifier,
  Label = NULL,
  Creator = NULL,
  Publisher = NULL,
  Issued = NULL,
  Subject = NULL
)

Arguments

x

A data.frame or inherited tibble, data.frame, or a structured list.

Dimensions

The name or column number of the dimensions within the dataset.

Measures

The name or column number of the measures within the dataset.

Attributes

The name or column number of the attributes within the dataset.

Title

dct:title, a name given to the resource. datacite allows the use of alternate titles, too. See dataset_title.

Identifier

An unambiguous reference to the resource within a given context. Recommended practice is to identify the resource by means of a string conforming to an identification system. Examples include International Standard Book Number (ISBN), Digital Object Identifier (DOI), and Uniform Resource Name (URN). Select and identifier scheme from registered URI schemes maintained by IANA. More details: Guidelines for using resource identifiers in Dublin Core metadata and IEEE LOM. Similar to Identifier in datacite. See identifier.

Label

may be used to provide a human-readable version of the dataset's name. A text description (optionally with a language tag) as defined by rdfs:label.

Creator

An entity primarily responsible for making the resource. dct:creator Corresponds to Creator in datacite. See creator.

Publisher

Corresponds to dct:publisher and Publisher in DataCite. The name of the entity that holds, archives, publishes prints, distributes, releases, issues, or produces the resource. This property will be used to formulate the citation, so consider the prominence of the role. For software, use Publisher for the code repository. If there is an entity other than a code repository, that "holds, archives, publishes, prints, distributes, releases, issues, or produces" the code, use the property Contributor/contributorType/hostingInstitution for the code repository. See publisher.

Issued

Corresponds to dct:date.

Subject

Recommended for discovery in DataCite. Subject, keyword, classification code, or key phrase describing the resource. Similar to dct:subject.
Use subject to properly add a key phrase from a controlled vocabulary and create structured Subject objects with subject_create.

Value

A well-described and formatted data frame with survey data and rich metadata.

Examples

path <- system.file("ZA5688_v6_sample.sav", package = "surveydataset")
survey_raw       <- read_spss_declared(path)

Identifier <- survey_raw$doi[1]

y <- survey(
  x = survey_raw,
  Dimensions = c(10, 22:33),
  Measures = c(13:21),
  Attributes = c(7:9, 11),
  Constants = c(1:6),
  Title = "Test Survey",
  Identifier = survey_raw$doi[1],
  Label = as.character(as.factor(survey_raw$survey[1]))
)